testing

+++ src/testing/ExampleSingleFFT.cu
@@ -6,6 +6,8 @@
 #include <memory>
 #include <cassert>
 
+#include <stdint.h>
+
 #include <cuda_fp16.h>
 #include <cuda_runtime.h>
 #include <cuda.h>
@@ -38,8 +40,8 @@ int main(){
 
   //The plan holds parameters needed for the execution of the kernels which are
   //mostly derived from the fft length.
-  std::optional<Plan<long long>> possible_plan = CreatePlan(fft_length);
-  Plan<long long> my_plan;
+  std::optional<Plan<int64_t>> possible_plan = CreatePlan(fft_length);
+  Plan<int64_t> my_plan;
   if (possible_plan) {
     my_plan = possible_plan.value();
   } else {
